Improve performance by dividing up tasks
The stream of inkscape commands have been divided across multiple
instances of inkscape, one per core. This speeds up rendering by
using more cpu cores, even for single svg rendering. To do this a
simple java program was written which starts N instances of
whatever arguments was passed to it and then reads one line at the
time from standard input, the lines are then divided into N
seperate streams which are then piped to the the difference instances
of whatever program was passed as argument to SplitJob.