Add filtered rendering
Instead of rendering out everything in an svg-file,
the render script now takes in a regular expression that is uses to
search all the object-IDs inside each of the svg-files for a match
and only renders those. To make the search go faster all object-IDs
are stored in files named ".<svg-file-name>.cache" after every full
render of an svg-file or if none exist.